added `dataAccess.js`; removed constituent files
`dataAccess.js` is a merger of the two earlier data access layer files.
I've decided to streamline things, both because the resultant file is
easier to submit for my class and because these functions are all pretty
straightforward. If it gets unwieldy in the future, I'll revisit this, I
suppose, but for right now, I think this is the best solution.

The two files that constitute `dataAccess.js` have also been removed.